+++ include/secrets_int.h
@@ -1,0 +1,8 @@
+
+struct fld {
+    const char *name;
+    size_t offset;
+};
+
+extern const struct fld RSA_private_field[];
+extern const int RSA_private_field_count;
