+++ lib/liboswkeys/rsapub.c
@@ -1,0 +1,153 @@
+/*
+ * mechanisms for preshared keys (public, private, and preshared secrets)
+ * this is the library for reading (and later, writing!) the ipsec.secrets
+ * files.
+ *
+ * Copyright (C) 1998-2004  D. Hugh Redelmeier.
+ * Copyright (C) 2017 Michael Richardson <mcr@xelerance.com>
+ */
+
+#include <stddef.h>
+#include <stdlib.h>
+#include <string.h>
+#include <ctype.h>
+#include <unistd.h>
+#include <errno.h>
+#include <time.h>
+#include <sys/socket.h>
+#include <netinet/in.h>
+#include <arpa/inet.h>
+#include <arpa/nameser.h>	/* missing from <resolv.h> on old systems */
+
+#include <gmp.h>
+#include <openswan.h>
+#include <openswan/ipsec_policy.h>
+
+#include "sysdep.h"
+#include "oswlog.h"
+#include "constants.h"
+#include "oswalloc.h"
+#include "oswtime.h"
+#include "id.h"
+#include "secrets_int.h"
+#include "secrets.h"
+#include "mpzfuncs.h"
+
+
+#ifdef DEBUG
+static void
+RSA_show_key_fields(struct RSA_private_key *k, int fieldcnt)
+{
+    const struct fld *p;
+
+    DBG_log(" keyid: *%s", k->pub.keyid);
+
+    for (p = RSA_private_field; p < &RSA_private_field[fieldcnt]; p++)
+    {
+	MP_INT *n = (MP_INT *) ((char *)k + p->offset);
+	size_t sz = mpz_sizeinbase(n, 16);
+	char buf[RSA_MAX_OCTETS * 2 + 2];	/* ought to be big enough */
+
+	passert(sz <= sizeof(buf));
+	mpz_get_str(buf, 16, n);
+
+	DBG_log(" %s: %s", p->name, buf);
+    }
+}
+
+/* debugging info that compromises security! */
+#if 0
+static void
+RSA_show_private_key(struct RSA_private_key *k)
+{
+    RSA_show_key_fields(k, RSA_private_field_count);
+}
+#endif
+
+static void
+RSA_show_public_key(struct RSA_public_key *k)
+{
+    /* Kludge: pretend that it is a private key, but only display the
+     * first two fields (which are the public key).
+     */
+    passert(offsetof(struct RSA_private_key, pub) == 0);
+    RSA_show_key_fields((struct RSA_private_key *)k, 2);
+}
+#endif
+
+
+/* decode of RSA pubkey chunk
+ * - format specified in RFC 2537 RSA/MD5 Keys and SIGs in the DNS
+ * - exponent length in bytes (1 or 3 octets)
+ *   + 1 byte if in [1, 255]
+ *   + otherwise 0x00 followed by 2 bytes of length
+ * - exponent
+ * - modulus
+ */
+err_t
+unpack_RSA_public_key(struct RSA_public_key *rsa, const chunk_t *pubkey)
+{
+    chunk_t exponent;
+    chunk_t mod;
+
+    rsa->keyid[0] = '\0';	/* in case of keybolbtoid failure */
+
+    if (pubkey->len < 3)
+	return "RSA public key blob way to short";	/* not even room for length! */
+
+    rsa->key_rfc3110 = chunk_clone(*pubkey, "rfc3110 format of public key");
+
+    if (pubkey->ptr[0] != 0x00)
+    {
+	setchunk(exponent, pubkey->ptr + 1, pubkey->ptr[0]);
+    }
+    else
+    {
+	setchunk(exponent, pubkey->ptr + 3
+	    , (pubkey->ptr[1] << BITS_PER_BYTE) + pubkey->ptr[2]);
+    }
+
+    if (pubkey->len - (exponent.ptr - pubkey->ptr) < exponent.len + RSA_MIN_OCTETS_RFC)
+	return "RSA public key blob too short";
+
+    mod.ptr = exponent.ptr + exponent.len;
+    mod.len = &pubkey->ptr[pubkey->len] - mod.ptr;
+
+    if (mod.len < RSA_MIN_OCTETS)
+	return RSA_MIN_OCTETS_UGH;
+
+    if (mod.len > RSA_MAX_OCTETS)
+	return RSA_MAX_OCTETS_UGH;
+
+    if (mod.len > pubkey->ptr + pubkey->len - mod.ptr)
+       return "RSA public key blob too short";
+
+    n_to_mpz(&rsa->e, exponent.ptr, exponent.len);
+    n_to_mpz(&rsa->n, mod.ptr, mod.len);
+
+    keyblobtoid(pubkey->ptr, pubkey->len, rsa->keyid, sizeof(rsa->keyid));
+
+#ifdef DEBUG
+    DBG(DBG_PRIVATE, RSA_show_public_key(rsa));
+#endif
+
+
+    rsa->k = mpz_sizeinbase(&rsa->n, 2);	/* size in bits, for a start */
+    rsa->k = (rsa->k + BITS_PER_BYTE - 1) / BITS_PER_BYTE;	/* now octets */
+
+    if (rsa->k != mod.len)
+    {
+	mpz_clear(&rsa->e);
+	mpz_clear(&rsa->n);
+	return "RSA modulus shorter than specified";
+    }
+
+    return NULL;
+}
+
+/*
+ * Local Variables:
+ * c-basic-offset:4
+ * c-style: pluto
+ * End:
+ */
