+++ tests/unit/libpluto/lp24-certreplydaveR2/output1.txt
@@ -839,9 +839,6 @@ sending 376 bytes for STATE_IKEv2_START through eth0:500 to 93.184.216.34:500 (u
 | reached self-signed root ca
 ./certreplydaveR2 0: validated X509 certificate: 'C=CA, ST=Ontario, L=Ottawa, O=Xelerance Corporation, OU=Doctor Taylor Plumage, CN=carol.testing.openswan.org, E=mcr@xelerance.com', added to trusted database
 | Public key validated
-|  keyid: *AwEAAbfz1
-|  Modulus: b7f3d43717cadfec3802e2d4867f45cb8f6b99bee1f892493db05565ab970de15237a2c0b7e0a0baed29a197e22f80b7cd99a25da675b1171693bef6ac579623f6bb47dec6a3f8981b99bacdc42c0f3610e2f4950d9a8408bc3b7a9a7e40cc05282fa1e26913035ed34520879de0ee14fb770f2d4a070f1aa4245aeb4820a4fd4fa7bf1db1cd8d1b58970f59947b29f204988e0fd0eaf57aa1fb6db9e3140fd351bfcc3663beab9d14665067f71a5345166831667b42251b9896d9f157b55c085a29a726ae62dfdf3e68c0c5b3d1ba11ca69c9f3233abcdc8c5fe5f4a3313965b8609f5855ff34610d5da981a8286444b1e6b821cd7b1c51a7db361754b5d211
-|  PublicExponent: 10001
 | ikev2 verify required CA is 'C=CA, ST=Ontario, L=Ottawa, O=Xelerance Corporation, OU=Doctor Taylor Plumage, CN=testing.xelerance.com, E=mcr@xelerance.com'
 |
 | List of X.509 End Certificates:
@@ -2298,9 +2295,6 @@ sending 376 bytes for STATE_IKEv2_START through eth0:500 to 93.184.216.35:500 (u
 | reached self-signed root ca
 ./certreplydaveR2 0: validated X509 certificate: 'C=CA, ST=ON, L=Ottawa, O=Xelerance Corporation, OU=Openswan, CN=dave@openswan.org, E=mcr@xelerance.com', added to trusted database
 | Public key validated
-|  keyid: *AwEAAeOoT
-|  Modulus: e3a84cc6ccc7e3b77ce74f154b04cafacb0646c0dc08c1c8aa5d06b77c5541abcfa800a9f38a50daa4c42c3e45e0f3b791282fd3b889cfe2c2339a9d04873c9cc3cd947f51703c7fda9b6d5e411db36c9e8476c2d49721fb4b869b103b437442942de0bd2bd7740e70993ac106f2b645ab0ccc1ebf90368008e0d8ed10c7a6e3df261d9573dbe6f0506c17b86e77fc54c9211eaf1a2b4a5dd3b07b5dc590b801141085bc86c25e4573d8e7a647605cee8580f449b3dfe28d50b6d5baf87909535501315716ecbce277480fe2c84ced69a9d8f1810eeac33e7d24c8bf7505882de7edc5cb29ceecef5785f9f47f37c7a7b87947ffea685ca873dc80e7b9c7ec67
-|  PublicExponent: 10001
 | ikev2 verify required CA is 'C=CA, ST=Ontario, L=Ottawa, O=Xelerance Corporation, OU=Doctor Taylor Plumage, CN=testing.xelerance.com, E=mcr@xelerance.com'
 |
 | List of X.509 End Certificates:
